* Added spawn method in WorldManager. * Added script functions 'Spawn', 'Rnd', and 'Notice'. * Changed gm command 'spawn', to spawn creatures in a specific radius around the player. * Added "doc/script_samples/", for longer examples.

+++ src/Common/Data/Spawns.cs
@@ -134,7 +134,7 @@ namespace Common.Data
 		public Line SpawnLine;
 		public SpawnRegion SpawnPolyRegion;
 		public RectangleF SpawnPolyBounds;
-		public byte Amount;
+		public uint Amount;
 
 		public MabiVertex GetRandomSpawnPoint()
 		{
@@ -165,7 +165,6 @@ namespace Common.Data
 
 			return new MabiVertex(x, y);
 		}
-
 	}
 
 	public class SpawnDb : DataManager<SpawnInfo>
